@@ -3,13 +3,20 @@
 	"version": "3.3.2",
 	"description": "Finds the common standard cache directory",
 	"license": "MIT",
-	"repository": "avajs/find-cache-dir",
-	"funding": "https://github.com/avajs/find-cache-dir?sponsor=1",
+	"repository": "sindresorhus/find-cache-dir",
+	"funding": "https://github.com/sponsors/sindresorhus",
+	"author": {
+		"name": "Sindre Sorhus",
+		"email": "sindresorhus@gmail.com",
+		"url": "https://sindresorhus.com"
+	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=8"
+		"node": ">=14.16"
 	},
 	"scripts": {
-		"test": "xo && nyc ava"
+		"test": "xo && ava"
 	},
 	"files": [
 		"index.js"
@@ -23,22 +30,16 @@
 		"search"
 	],
 	"dependencies": {
-		"commondir": "^1.0.1",
-		"make-dir": "^3.0.2",
-		"pkg-dir": "^4.1.0"
+		"common-path-prefix": "^3.0.0",
+		"pkg-dir": "^7.0.0"
 	},
 	"devDependencies": {
-		"ava": "^2.4.0",
-		"coveralls": "^3.0.9",
-		"del": "^4.0.0",
-		"nyc": "^15.0.0",
-		"tempy": "^0.4.0",
-		"xo": "^0.25.3"
+		"ava": "^5.0.1",
+		"del": "^7.0.0",
+		"tempy": "^3.0.0",
+		"xo": "^0.52.4"
 	},
-	"nyc": {
-		"reporter": [
-			"lcov",
-			"text"
-		]
+	"ava": {
+		"workerThreads": false
 	}
 }
