@@ -1,9 +1,8 @@
-'use strict';
-const path = require('path');
-const fs = require('fs');
-const commonDir = require('commondir');
-const pkgDir = require('pkg-dir');
-const makeDir = require('make-dir');
+import process from 'node:process';
+import path from 'node:path';
+import fs from 'node:fs';
+import commonPathPrefix from 'common-path-prefix';
+import {packageDirectorySync} from 'pkg-dir';
 
 const {env, cwd} = process;
 
@@ -11,14 +10,14 @@ const isWritable = path => {
 	try {
 		fs.accessSync(path, fs.constants.W_OK);
 		return true;
-	} catch (_) {
+	} catch {
 		return false;
 	}
 };
 
 function useDirectory(directory, options) {
 	if (options.create) {
-		makeDir.sync(directory);
+		fs.mkdirSync(directory, {recursive: true});
 	}
 
 	if (options.thunk) {
@@ -32,8 +31,8 @@ function getNodeModuleDirectory(directory) {
 	const nodeModules = path.join(directory, 'node_modules');
 
 	if (
-		!isWritable(nodeModules) &&
-		(fs.existsSync(nodeModules) || !isWritable(path.join(directory)))
+		!isWritable(nodeModules)
+			&& (fs.existsSync(nodeModules) || !isWritable(path.join(directory)))
 	) {
 		return;
 	}
@@ -41,7 +40,7 @@ function getNodeModuleDirectory(directory) {
 	return nodeModules;
 }
 
-module.exports = (options = {}) => {
+export default function findCacheDirectory(options = {}) {
 	if (env.CACHE_DIR && !['true', 'false', '1', '0'].includes(env.CACHE_DIR)) {
 		return useDirectory(path.join(env.CACHE_DIR, options.name), options);
 	}
@@ -49,10 +48,10 @@ module.exports = (options = {}) => {
 	let {cwd: directory = cwd()} = options;
 
 	if (options.files) {
-		directory = commonDir(directory, options.files);
+		directory = commonPathPrefix(options.files.map(file => path.resolve(directory, file)));
 	}
 
-	directory = pkgDir.sync(directory);
+	directory = packageDirectorySync({cwd: directory});
 
 	if (!directory) {
 		return;
@@ -60,8 +59,8 @@ module.exports = (options = {}) => {
 
 	const nodeModules = getNodeModuleDirectory(directory);
 	if (!nodeModules) {
-		return undefined;
+		return;
 	}
 
 	return useDirectory(path.join(directory, 'node_modules', '.cache', options.name), options);
-};
+}
